s390: make cpu_relax() a memory barrier
Change cpu_relax() so that is a memory barrier which generates the bcr
15/14 instruction and therefore enforces serialization.

This makes sure that e.g. a loop like

     while (atomic_read(&val))
     	   cpu_relax();

will see changes to val fast. Otherwise and in theory according to the
architecture, the while loop would be allowed to never finish, even if
a different cpu changed the value of val to zero.

Acked-by: Martin Schwidefsky <schwidefsky@de.ibm.com>
Signed-off-by: Heiko Carstens <heiko.carstens@de.ibm.com>
Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>