s390/locking: add additional memory barriers
Add serializing instructions (bcr 14/15) behind the instruction that
unlocks spin_locks and before instructions that fetch the lock value
in a loop to test if a lock is unlocked.

Otherwise and in theory according to the architecture, those loops
would be allowed to never finish, even if a different cpu unlocked a
lock.

Signed-off-by: Heiko Carstens <heiko.carstens@de.ibm.com>
Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>