ipmi: allow dynamic BMC version information
Currently, it's up to the IPMI SMIs to provide the product & version
details of BMCs behind registered IPMI SMI interfaces. This device ID is
provided on SMI regsitration, and kept around for all future queries.

However, this version information isn't always static. For example, a
BMC may be upgraded at runtime, making the old version information
stale.

This change allows querying the BMC device ID & version information
dynamically. If no static device_id argument is provided to
ipmi_register_smi, then the IPMI core code will perform a Get Device ID
IPMI command to query the version information when needed. We keep a
short-term cache of this information so we don't need to re-query
for every attribute access.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>

Changed the dyn_id element to dyn_id_state to better reflect its
function, and defined it using the enum that it represents.

Added a comment above null_user_handler to state that it was
used by querying the BMC device ID, and if it was used by anything
else a mutex would be required.

Signed-off-by: Corey Minyard <cminyard@mvista.com>