ARM64: dts: meson: Work around for new secmon zones
Depending the platform and u-boot binary used the ATF might be using
different memory zones.

Until now, we assumed that it was always 0x10000000-0x10200000, this
was clearly wrong. Accessing the ATF memory result in "Synchronous
External Abort" error in the kernel

While we need to come up with a better solution. let's work around the
issue and protect what the zone is know to be on recent libretech
releases

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>